Added server side group control

+++ server.lua
@@ -19,7 +19,6 @@ ESX.RegisterServerCallback("hxz:checkgroup", function(source, cb)
     end
 end)
 
-
 ESX.RegisterServerCallback('hxz:checkTarget', function(source, cb, target)
     local xTarget = ESX.GetPlayerFromId(target)
     if xTarget then
@@ -30,46 +29,56 @@ end)
 RegisterNetEvent("hxz:bullet")
 AddEventHandler("hxz:bullet", function(id)
     local xPlayer = ESX.GetPlayerFromId(id)
-    if xPlayer == nil then
-        TriggerClientEvent('esx:showNotification', source, 'Player non trovato')
+    local Trigger = ESX.GetPlayerFromId(source)
+
+    if Permission(Trigger) then
+        if xPlayer == nil then
+            TriggerClientEvent('esx:showNotification', source, 'Player non trovato')
+        else
+            SetPedArmour(xPlayer.source, 100)
+        end
     else
-        SetPedArmour(xPlayer.source, 100)
+        DropPlayer(source, Lang['NOTIFY_DROP_TRIGGER'])
     end
 end)
 
-RegisterServerEvent('hxz:wipepg')
+RegisterNetEvent('hxz:wipepg')
 AddEventHandler('hxz:wipepg', function(id)
-
     local xPlayer = ESX.GetPlayerFromId(id)
-    if xPlayer == nil then
-        TriggerClientEvent('esx:showNotification', source, Lang['notify_player_not_found'])
-    else
-        DropPlayer(id, Lang['drop_notify'])
+    local Trigger = ESX.GetPlayerFromId(source)
 
-        MySQL.Sync.execute('DELETE FROM billing WHERE identifier = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM datastore_data WHERE owner = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM owned_properties WHERE owner = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM owned_vehicles WHERE owner = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM users WHERE identifier = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM user_licenses WHERE owner = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM multicharacter_slots WHERE identifier = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
-        MySQL.Sync.execute('DELETE FROM tattoos WHERE identifier = @identifier', {
-            ['@identifier'] = xPlayer.identifier
-        })
+    if Permission(Trigger) then
+        if xPlayer == nil then
+            TriggerClientEvent('esx:showNotification', source, Lang['NOTIFY_PLAYER_NOT_FOUND'])
+        else
+            DropPlayer(id, Lang['NOTIFY_DROP_PLAYER'])
+            MySQL.Sync.execute('DELETE FROM billing WHERE identifier = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM datastore_data WHERE owner = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM owned_properties WHERE owner = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM owned_vehicles WHERE owner = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM users WHERE identifier = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM user_licenses WHERE owner = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM multicharacter_slots WHERE identifier = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+            MySQL.Sync.execute('DELETE FROM tattoos WHERE identifier = @identifier', {
+                ['@identifier'] = xPlayer.identifier
+            })
+        end
+    else
+        DropPlayer(source, Lang['NOTIFY_DROP_TRIGGER'])
     end
 end)
 
@@ -77,16 +86,31 @@ RegisterServerEvent('hxz:setVehicle')
 AddEventHandler('hxz:setVehicle', function (vehicleProps, playerID, VehicleType, plate)
 	local _source = playerID
 	local xPlayer = ESX.GetPlayerFromId(_source)
+    local Trigger = ESX.GetPlayerFromId(source)
 
-	MySQL.Sync.execute('INSERT INTO owned_vehicles (owner, plate, vehicle, stored, type) VALUES (@owner, @plate, @vehicle, @stored, @type)',
-	{
-		['@owner']   = xPlayer.identifier,
-		['@plate']   = vehicleProps.plate,
-		['@vehicle'] = json.encode(vehicleProps),
-		['@stored']  = 1,
-		['@state']   = 0,
-		['type'] = VehicleType
-	}, function ()
-		TriggerClientEvent('esx:showNotification', _source, 'Hai ricevuto un veicolo')
-	end)
+
+    if Permission(Trigger) then
+        MySQL.Sync.execute('INSERT INTO owned_vehicles (owner, plate, vehicle, stored, type) VALUES (@owner, @plate, @vehicle, @stored, @type)',
+        {
+            ['@owner']   = xPlayer.identifier,
+            ['@plate']   = vehicleProps.plate,
+            ['@vehicle'] = json.encode(vehicleProps),
+            ['@stored']  = 1,
+            ['@state']   = 0,
+            ['type'] = VehicleType
+        }, function ()
+        end)
+    else
+        DropPlayer(source, Lang['NOTIFY_DROP_TRIGGER'])
+    end
 end)
+
+
+function Permission(Trigger)
+	for k,v in ipairs(Config.Staff) do
+		if Trigger.getGroup() == v then 
+			return true 
+		end
+	end
+	return false
+end
